pyproject.toml
[tool.poetry]
name = "quotes-dbx"
version = "0.1.0"
description = ""
authors = ["doug-pires <dpiresmartins@yahoo.com.br>"]
readme = "README.md"
packages = [{include = "quotes_dbx"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
dbx = "0.8.18"
pytest = "^7.4.2"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["--strict-markers", "--strict-config", "-ra"]
testpaths = ["tests"]

[tool.black]
line-length = 88